+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/project/BeanTransformerConfiguration.java
@@ -1,10 +1,14 @@
 package com.dwarfeng.familyhelper.webapi.node.configuration.project;
 
 import com.dwarfeng.familyhelper.project.sdk.bean.entity.FastJsonTaskTypeIndicator;
+import com.dwarfeng.familyhelper.project.sdk.bean.entity.JSFixedFastJsonPop;
 import com.dwarfeng.familyhelper.project.sdk.bean.entity.JSFixedFastJsonProject;
+import com.dwarfeng.familyhelper.project.stack.bean.entity.Pop;
 import com.dwarfeng.familyhelper.project.stack.bean.entity.Project;
 import com.dwarfeng.familyhelper.project.stack.bean.entity.TaskTypeIndicator;
+import com.dwarfeng.familyhelper.webapi.sdk.bean.disp.project.JSFixedFastJsonDispPop;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.disp.project.JSFixedFastJsonDispProject;
+import com.dwarfeng.familyhelper.webapi.stack.bean.disp.project.DispPop;
 import com.dwarfeng.familyhelper.webapi.stack.bean.disp.project.DispProject;
 import com.dwarfeng.subgrade.impl.bean.DozerBeanTransformer;
 import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
@@ -29,6 +33,16 @@ public class BeanTransformerConfiguration {
     @Bean
     public BeanTransformer<DispProject, JSFixedFastJsonDispProject> dispProjectBeanTransformer() {
         return new DozerBeanTransformer<>(DispProject.class, JSFixedFastJsonDispProject.class, mapper);
+    }
+
+    @Bean
+    public BeanTransformer<Pop, JSFixedFastJsonPop> popBeanTransformer() {
+        return new DozerBeanTransformer<>(Pop.class, JSFixedFastJsonPop.class, mapper);
+    }
+
+    @Bean
+    public BeanTransformer<DispPop, JSFixedFastJsonDispPop> dispPopBeanTransformer() {
+        return new DozerBeanTransformer<>(DispPop.class, JSFixedFastJsonDispPop.class, mapper);
     }
 
     @Bean